Moved limit switch to separate file
